move files into folders for organization
